Extract ForgedTransactionsVerifier
