+++ packages/processor/source/block-processor.ts
@@ -3,7 +3,13 @@ import { Contracts, Identifiers } from "@mainsail/contracts";
 import { Services, Utils as AppUtils } from "@mainsail/kernel";
 
 import { AcceptBlockHandler } from "./handlers";
-import { ChainedVerifier, IncompatibleTransactionsVerifier, NonceVerifier, VerifyBlockVerifier } from "./verifiers";
+import {
+	ChainedVerifier,
+	ForgedTransactionsVerifier,
+	IncompatibleTransactionsVerifier,
+	NonceVerifier,
+	VerifyBlockVerifier,
+} from "./verifiers";
 
 @injectable()
 export class BlockProcessor implements Contracts.BlockProcessor.Processor {
@@ -12,12 +18,6 @@ export class BlockProcessor implements Contracts.BlockProcessor.Processor {
 
 	@inject(Identifiers.LogService)
 	private readonly logger!: Contracts.Kernel.Logger;
-
-	@inject(Identifiers.Database.Service)
-	private readonly databaseService: Contracts.Database.IDatabaseService;
-
-	@inject(Identifiers.StateStore)
-	private readonly stateStore!: Contracts.State.StateStore;
 
 	@inject(Identifiers.TriggerService)
 	private readonly triggers!: Services.Triggers.Triggers;
@@ -46,53 +46,11 @@ export class BlockProcessor implements Contracts.BlockProcessor.Processor {
 		// 	return this.app.resolve<InvalidGeneratorHandler>(InvalidGeneratorHandler).execute(block);
 		// }
 
-		const containsForgedTransactions: boolean = await this.#checkBlockContainsForgedTransactions(roundState);
-		if (containsForgedTransactions) {
+		if (!(await this.app.resolve(ForgedTransactionsVerifier).execute(roundState))) {
 			return false;
 		}
 
 		return this.app.resolve<AcceptBlockHandler>(AcceptBlockHandler).execute(roundState);
-	}
-
-	async #checkBlockContainsForgedTransactions(roundState: Contracts.Consensus.IRoundState): Promise<boolean> {
-		const block = roundState.getProposal().toData().block;
-		if (block.transactions.length > 0) {
-			const transactionIds = block.transactions.map((tx) => {
-				AppUtils.assert.defined<string>(tx.id);
-
-				return tx.id;
-			});
-
-			const forgedIds: string[] = await this.databaseService.getForgedTransactionsIds(transactionIds);
-
-			if (this.stateStore.getLastBlock().data.height !== this.stateStore.getLastStoredBlockHeight()) {
-				const transactionIdsSet = new Set<string>(transactionIds);
-
-				for (const stateBlock of this.stateStore
-					.getLastBlocks()
-					.filter((block) => block.data.height > this.stateStore.getLastStoredBlockHeight())) {
-					for (const tx of stateBlock.transactions) {
-						AppUtils.assert.defined<string>(tx.id);
-
-						if (transactionIdsSet.has(tx.id)) {
-							forgedIds.push(tx.id);
-						}
-					}
-				}
-			}
-
-			if (forgedIds.length > 0) {
-				this.logger.warning(
-					`Block ${block.data.height.toLocaleString()} disregarded, because it contains already forged transactions`,
-				);
-
-				this.logger.debug(`${JSON.stringify(forgedIds, undefined, 4)}`);
-
-				return true;
-			}
-		}
-
-		return false;
 	}
 
 	// @ts-ignore
