+++ packages/processor/source/verifiers/index.ts
@@ -1,4 +1,5 @@
 export * from "./chained-verifier";
+export * from "./forged-transactions-verifier";
 export * from "./incompatible-transactions-verifier";
 export * from "./nonce-verifier";
 export * from "./verify-block-verifier";
